requirements.txt
absl-py==0.8.1
astor==0.8.0
Django==2.2.6
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.2
opt-einsum==3.1.0
pandas==0.25.1
pathlib==1.0.1
Pillow==6.2.0
protobuf==3.10.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
